refactor hub manager out of main
