+++ sample-solutions/VisionSampleToolset/IotEdgeSolutionNew/modules/VisionSampleModule/module.json
@@ -4,7 +4,7 @@
   "image": {
     "repository": "$CONTAINER_REGISTRY_NAME/visionsamplemodule",
     "tag": {
-      "version": "0.2.89",
+      "version": "0.2.99",
       "platforms": {
         "amd64": "./Dockerfile.amd64",
         "amd64.debug": "./Dockerfile.amd64.debug",
